backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pillow==10.1.0
<<<<<<< HEAD

# PyTorch (ARM64 Linux compatible)
torch==2.2.2
torchvision==0.17.2

# Transformers + NLP
transformers==4.35.2
sentencepiece==0.2.0
safetensors==0.6.2

# Critical fix: numpy must be <2.0 for scikit-learn compatibility
numpy==1.26.4

# scikit-learn version compatible with numpy<2
scikit-learn==1.3.2

=======
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
sentencepiece==0.1.99
numpy==1.24.3
scikit-learn==1.3.2
>>>>>>> 7317999ac0186241bdad8633188701b09657ab9f
boto3==1.29.7
python-dotenv==1.0.0
aiofiles==23.2.1
python-json-logger==2.0.7
<<<<<<< HEAD
asyncpg==0.29.0
motor==3.3.2
=======

>>>>>>> 7317999ac0186241bdad8633188701b09657ab9f
